chore: check if address link exists before adding in projections

+++ src/ParcelRegistry.Projections.Integration/ParcelVersion/ParcelVersionProjections.cs
@@ -9,6 +9,7 @@
     using Converters;
     using Infrastructure;
     using Legacy.Events;
+    using Microsoft.EntityFrameworkCore;
     using Microsoft.Extensions.Options;
     using Parcel;
     using Parcel.Events;
@@ -114,13 +115,12 @@
                     message,
                     _ => { }, ct);
 
-                await context
-                    .ParcelVersionAddresses
-                    .AddAsync(new ParcelVersionAddress(
-                        message.Position,
-                        message.Message.ParcelId,
-                        message.Message.AddressPersistentLocalId,
-                        message.Message.CaPaKey), ct);
+                await AddParcelAddress(context,
+                    message.Position,
+                    message.Message.ParcelId,
+                    message.Message.CaPaKey,
+                    message.Message.AddressPersistentLocalId,
+                    ct);
             });
 
             When<Envelope<ParcelAddressWasReplacedBecauseOfMunicipalityMerger>>(async (context, message, ct) =>
@@ -136,13 +136,12 @@
                     message.Message.PreviousAddressPersistentLocalId,
                     ct);
 
-                await context
-                    .ParcelVersionAddresses
-                    .AddAsync(new ParcelVersionAddress(
-                        message.Position,
-                        message.Message.ParcelId,
-                        message.Message.NewAddressPersistentLocalId,
-                        message.Message.CaPaKey), ct);
+                await AddParcelAddress(context,
+                    message.Position,
+                    message.Message.ParcelId,
+                    message.Message.CaPaKey,
+                    message.Message.NewAddressPersistentLocalId,
+                    ct);
             });
 
             When<Envelope<ParcelAddressWasReplacedBecauseAddressWasReaddressed>>(async (context, message, ct) =>
@@ -169,7 +168,7 @@
                     .ParcelVersionAddresses
                     .FindAsync([message.Position, message.Message.ParcelId, message.Message.NewAddressPersistentLocalId], cancellationToken: ct);
 
-                if (newAddress is null)
+                if (newAddress is null || context.Entry(newAddress).State == EntityState.Deleted)
                 {
                     await context
                         .ParcelVersionAddresses
@@ -194,32 +193,21 @@
 
                 foreach (var addressPersistentLocalId in message.Message.DetachedAddressPersistentLocalIds)
                 {
-                    var relation = await context
-                        .ParcelVersionAddresses
-                        .FindAsync([message.Position, message.Message.ParcelId, addressPersistentLocalId], ct);
-
-                    if (relation is not null)
-                    {
-                        context.ParcelVersionAddresses.Remove(relation);
-                    }
+                    await RemoveParcelAddress(context,
+                        message.Position,
+                        message.Message.ParcelId,
+                        addressPersistentLocalId,
+                        ct);
                 }
 
                 foreach (var addressPersistentLocalId in message.Message.AttachedAddressPersistentLocalIds)
                 {
-                    var relation = await context
-                        .ParcelVersionAddresses
-                        .FindAsync([message.Position, message.Message.ParcelId, addressPersistentLocalId], ct);
-
-                    if (relation is null)
-                    {
-                        await context.ParcelVersionAddresses.AddAsync(
-                            new ParcelVersionAddress(
-                                message.Position,
-                                message.Message.ParcelId,
-                                addressPersistentLocalId,
-                                message.Message.CaPaKey),
-                            ct);
-                    }
+                    await AddParcelAddress(context,
+                        message.Position,
+                        message.Message.ParcelId,
+                        message.Message.CaPaKey,
+                        addressPersistentLocalId,
+                        ct);
                 }
             });
 
@@ -455,6 +443,30 @@
 
         #endregion
 
+        private static async Task AddParcelAddress(
+            IntegrationContext context,
+            long position,
+            Guid parcelId,
+            string caPaKey,
+            int addressPersistentLocalId,
+            CancellationToken ct)
+        {
+            var newAddress = await context
+                .ParcelVersionAddresses
+                .FindAsync([position, parcelId, addressPersistentLocalId], cancellationToken: ct);
+
+            if (newAddress is null || context.Entry(newAddress).State == EntityState.Deleted)
+            {
+                await context
+                    .ParcelVersionAddresses
+                    .AddAsync(new ParcelVersionAddress(
+                        position,
+                        parcelId,
+                        addressPersistentLocalId,
+                        caPaKey), ct);
+            }
+        }
+
         private static async Task RemoveParcelAddress(
             IntegrationContext context,
             long position,
@@ -467,7 +479,10 @@
                 .FindAsync(new object?[] { position, parcelId, addressPersistentLocalId },
                     cancellationToken: ct);
 
-            context.ParcelVersionAddresses.Remove(versionAddress);
+            if (versionAddress is not null)
+            {
+                context.ParcelVersionAddresses.Remove(versionAddress);
+            }
         }
     }
 }
